chore: reorganize core decoration tests
